+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Changelog for the gruf gem. This includes internal history before the gem was ma
 ### Pending release
 
 - Drop support for Ruby 2.4/2.5 to align with Ruby EOL schedule, supporting 2.6+ only
+- Allow for float/TimeSpec timeout values on clients
 
 ### 2.9.0